backend/python-api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
pydantic==2.11.7
email-validator==2.2.0
python-dotenv==1.1.1
pymysql==1.1.1
twilio==9.3.0
